Spawn apps and menus on the correct display
Spawn applications and menus on the display and screen where the pointer
resides.  Prior to starting an application via menu or binding, set the
DISPLAY environment variable to the display and screen where the pointer
is.  This is necessary in cases where there are multiple displays,
otherwise, all applications start on the first display.

If the DISPLAY variable cannot be parsed (i.e., missing colon), spawn
without touching DISPLAY.